Creates JSON file of correct type
Creates JSON file that is a dict with noun keys containing arrays of
verb objects. This program encorporates the use of tokenized
sentences and has one parsing method, which can be optimized by fusing
it with insert_verb_object. However, I forgot to use noun/verb filters,
but the two functions are present in createjson.py file